requirements.txt
requests==2.25.1
pytz==2021.1
python-telegram-bot==12.2.0